support single step test execution
