+++ fusion.js
@@ -1,11 +1,8 @@
 #!/usr/bin/env node
 'use strict';
 
+require('shelljs/global');
 const program = require('commander');
-const utils = require('./src/lib/utils');
-const FusionGenerator = require('./src/lib/fusionGenerator');
-const fs = require('fs');
-const co = require('co');
 
 program
   .version('1.0.0')
@@ -24,36 +21,5 @@ if (!folder || !testEnv || !browser) {
   process.exit(1);
 }
 
-let configFile = `/config/${testEnv}.yaml`;
-let config = utils.loadYAMLOrParse(folder, configFile);
-
-config.basePath = folder;
-config.browser = browser;
-
-let testPath = __dirname + '/test';
-
-config.basePath = folder;
-
-let requestFolders = config.requestFolders || ['requests'];
-
-for (let reqFolder of requestFolders) {
-  let path = `${folder}/${reqFolder}`;
-  let files = fs.readdirSync(path);
-  for (let fileName of files) {
-    config.feature = fileName.replace('.yaml', '');
-    let requests = utils.loadYAMLOrParse(path, fileName);
-    let fusionGeneratorInstance = new FusionGenerator(config, requests, path + '/' + fileName);
-    co(function *start() {
-      testPath += `/${config.feature}.js`;
-      yield fusionGeneratorInstance.fusion(testPath);
-    });
-  }
-}
-
-// Display uncaught Exception.
-process.on('uncaughtException', (err) => {
-  console.error((new Date()).toUTCString() + ' uncaughtException:', err.message);
-  console.error(err.stack);
-  process.exit(1);
-});
-
+exec(`generator -f ${folder} -e ${testEnv} -b ${browser}`, {silent:true});
+exec(`npm test`, {silent:false}).stdout;
