+++ package.json
@@ -8,7 +8,8 @@
     "lint": "./node_modules/.bin/eslint . && ./node_modules/.bin/jscs ."
   },
   "bin": {
-    "fusion": "./fusion.js"
+    "fusion": "./fusion.js",
+    "generator": "./generator.js"
   },
   "repository": {
     "type": "git",
@@ -33,6 +34,7 @@
     "lodash": "^4.17.2",
     "mkdirp": "^0.5.1",
     "selenium-webdriver": "^3.0.1",
+    "shelljs": "^0.7.5",
     "should": "^11.1.1",
     "zpad": "^0.5.0"
   },
